requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=25.1

numpy>=1.26,<2.1
pandas>=2.1,<2.4
pillow>=10.0

python-decouple>=3.8
requests>=2.31

psutil>=5.9
PyJWT>=2.8

asgiref>=3.7
sqlparse>=0.4
tzdata>=2024.1
